removed magical Segmenter class definition
Previous versions of the segmenter would define a Segmenter class inside of
the report class.  This meant that when adding segments, you would not have to
inherit from Dossier::Segmenter.  Now any report creating a segmenter, will
have to define the segmenter class as inheriting from Dossier::Segmenter.  As
the segmenter class no longer has a reference to the report class, all of the
Segment subclasses are created inside of the Segmenter namespace, instead of
the report namespace.